Migrates from EpicHoppers v4.7.8 to v5.0.0-SNAPSHOT(-tmp)
The -tmp suffix is because of a currently failing GitHub Action pipeline which cannot succeed
as long as specific other plugins EpicHoppers relies on have not been migrated to the new Core etc.

So I manually uploaded a *release* artifact to the repo (which will be deleted in the near future).